refactor(redux): move me actions with side effects to meThunks
